requirements.txt
fastapi>=0.110,<0.112
uvicorn[standard]>=0.29,<0.30
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0
supabase>=2.4,<3.0
httpx>=0.27,<0.28
PyJWT>=2.8,<3.0
slowapi>=0.1.8,<0.2
transformers>=4.39,<5.0
torch>=2.2,<3.0
numpy>=1.26,<2.0
sendgrid>=6.11,<7.0
celery>=5.3,<6.0
sentry-sdk>=2.10,<3.0
